Create DeleteTodoButton component
Fixes #68

Add `DeleteTodoButton` component and update `TodoList` to use it.

* **Add `DeleteTodoButton` component**: Create `src/components/DeleteTodoButton.tsx` with the provided code to handle the delete action for todos.
* **Update `TodoList` component**: Modify `src/components/TodoList.tsx` to remove the `useActionState` hook and related code, and import and use the new `DeleteTodoButton` component for the delete action.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/kenfj/nextjs15-todo-example/issues/68?shareId=XXXX-XXXX-XXXX-XXXX).
