+++ src/components/TodoList.tsx
@@ -1,25 +1,12 @@
 import { Todo } from '@prisma/client';
-import { useActionState } from 'react';
 
-import { deleteTodoAction } from '@/actions/todos';
-import { DeleteTodoState } from '@/models/todo';
+import DeleteTodoButton from '@/components/DeleteTodoButton';
 
 type TodoListProps = {
   todos?: Todo[];
 }
 
-const initialState: DeleteTodoState = {
-  success: false,
-  prismaError: "",
-};
-
 const TodoList = ({ todos }: TodoListProps) => {
-  const [state, formAction, pending] = useActionState(deleteTodoAction, initialState);
-
-  if (state.prismaError) {
-    return <div className="text-red-500">{state.prismaError}</div>;
-  }
-
   if (!todos || todos.length === 0) {
     return <div>No todos available</div>;
   }
@@ -29,12 +16,7 @@ const TodoList = ({ todos }: TodoListProps) => {
       {todos.map((todo) => (
         <li key={todo.id} className="flex items-center justify-between">
           {todo.title}
-          <form action={formAction}>
-            <input name="todoId" className="hidden" value={todo.id} readOnly />
-            <button type="submit" className="text-red-500 hover:text-red-700 ml-4" disabled={pending}>
-              &#x2716;
-            </button>
-          </form>
+          <DeleteTodoButton todoId={todo.id} />
         </li>
       ))}
     </ul>
